backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
flask-bcrypt==1.0.1
PyJWT==2.10.1
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
gunicorn==23.0.0
pytest==8.3.5
